updated to semistandard 11
updated the dependency and also updated the code syntax to conform to the new rules. A few of the spec files I had to disable the "no-unused-expressions" rule because otherwise the mocha chai syntax gives a false error "expect an assignment or function call and instead saw an expression". see https://stackoverflow.com/questions/45079454/no-unused-expressions-in-mocha-chai-unit-test-using-standardjs