Only load featurelayer tiles after metadata has been loaded
This can have a significant impact on performance for larger layers,
where the initial detailed tiles are requested concurrently with the
metadata request and thus loaded in arcgis format -- assuming isModern
hasn't been manually set -- and need to be converted to geojson (which
can choke a browser).

This is only a partial fix, as _setView can also trigger  _update.